vote and unvote comment

+++ code-assessment-service/code-assessment-domain/code-assessment-application-service/src/main/java/com/backend/programming/learning/system/code/assessment/service/domain/implement/service/shared_solution/comment/CommentHelper.java
@@ -2,11 +2,14 @@ package com.backend.programming.learning.system.code.assessment.service.domain.i
 
 import com.backend.programming.learning.system.code.assessment.service.domain.CodeAssessmentDomainService;
 import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.create.shared_solution.comment.CreateCommentCommand;
+import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.create.shared_solution.comment.vote.VoteCommentCommand;
 import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.delete.shared_solution.comment.DeleteCommentCommand;
+import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.delete.shared_solution.comment.vote.UnvoteCommentCommand;
 import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.query.shared_solution.comment.GetReplyCommentCommand;
 import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.query.shared_solution.comment.GetSolutionCommentCommand;
 import com.backend.programming.learning.system.code.assessment.service.domain.dto.method.update.shared_solution.comment.UpdateCommentCommand;
 import com.backend.programming.learning.system.code.assessment.service.domain.entity.Comment;
+import com.backend.programming.learning.system.code.assessment.service.domain.entity.CommentVote;
 import com.backend.programming.learning.system.code.assessment.service.domain.entity.SharedSolution;
 import com.backend.programming.learning.system.code.assessment.service.domain.entity.User;
 import com.backend.programming.learning.system.code.assessment.service.domain.exeption.CodeAssessmentDomainException;
@@ -16,6 +19,7 @@ import com.backend.programming.learning.system.code.assessment.service.domain.ma
 import com.backend.programming.learning.system.code.assessment.service.domain.ports.output.repository.CommentRepository;
 import com.backend.programming.learning.system.code.assessment.service.domain.valueobject.CommentId;
 import com.backend.programming.learning.system.code.assessment.service.domain.valueobject.SharedSolutionId;
+import com.backend.programming.learning.system.code.assessment.service.domain.valueobject.comment_vote.CommentVoteId;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -44,9 +48,13 @@ public class CommentHelper {
     public Comment createComment(CreateCommentCommand command) {
         User user = validateHelper.validateUser(command.getUserId());
 
-        validateHelper.validateSharedSolution(command.getSharedSolutionId());
+        SharedSolution sharedSolution = validateHelper.validateSharedSolution(command.getSharedSolutionId());
 
         Comment replyComment = command.getReplyId() == null? null: validateHelper.validateComment(command.getReplyId());
+
+        if(replyComment != null && !replyComment.getSharedSolutionId().equals(sharedSolution.getId())){
+            throw new CodeAssessmentDomainException("Shared solution " + command.getSharedSolutionId() + "does not own root comment " + command.getReplyId());
+        }
 
         Comment comment = commentDataMapper.createCommentCommandToCommand(command, user);
         codeAssessmentDomainService.intitateComment(comment, replyComment);
@@ -59,9 +67,12 @@ public class CommentHelper {
     public void updateComment(UpdateCommentCommand command) {
         User user = validateHelper.validateUser(command.getUserId());
 
-        validateHelper.validateSharedSolution(command.getSharedSolutionId());
+        SharedSolution sharedSolution = validateHelper.validateSharedSolution(command.getSharedSolutionId());
 
         Comment comment = validateHelper.validateComment(command.getCommentId());
+
+        if(!comment.getSharedSolutionId().equals(sharedSolution.getId()))
+            throw new CodeAssessmentDomainException("Shared solution " + command.getSharedSolutionId() + "does not own comment " + command.getCommentId());
 
         if(!comment.getUser().equals(user))
             throw new CodeAssessmentDomainException("User " + command.getUserId() + "does not own comment " + command.getCommentId());
@@ -74,9 +85,12 @@ public class CommentHelper {
 
     @Transactional
     public void deleteComment(DeleteCommentCommand command) {
-        validateHelper.validateUser(command.getUserId());
+        User user = validateHelper.validateUser(command.getUserId());
         validateHelper.validateSharedSolution(command.getSharedSolutionId());
-        validateHelper.validateComment(command.getCommentId());
+        Comment comment = validateHelper.validateComment(command.getCommentId());
+
+        if(!comment.getUser().equals(user))
+            throw new CodeAssessmentDomainException("User " + command.getUserId() + "does not own comment " + command.getCommentId());
 
         commentRepository.deleteById(new CommentId(command.getCommentId()));
     }
@@ -101,5 +115,22 @@ public class CommentHelper {
 
         return commentRepository.findReplyByRootCommentId(comment.getId(), user.getId());
 
+    }
+
+    @Transactional
+    public void voteComment(VoteCommentCommand command) {
+        validateHelper.validateUser(command.getUserId());
+        validateHelper.validateComment(command.getCommentId());
+        CommentVote commentVote = commentDataMapper.voteCommentCommandToCommentVote(command);
+        commentRepository.vote(commentVote);
+
+    }
+
+    @Transactional
+    public void unvoteComment(UnvoteCommentCommand command) {
+        CommentVoteId commentVoteId = commentDataMapper.unvoteCommentCommandToCommentVoteId(command);
+        validateHelper.validateCommentVote(commentVoteId);
+
+        commentRepository.unvote(commentVoteId);
     }
 }
